pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structor"
version = "1.0.3"
description = "Simple package that allow you to create templates of folders/files structure and generate them from a cli (like done in angular, django, react...)"
readme = "README.md"
authors = [{ name = "Mayas Nova" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["structor", "file generator"]

dependencies = [
    "click>=8.1.3",
    "colorama>=0.4.6",
    "commonmark>=0.9.1",
    "Pygments>=2.14.0",
    "PyYAML>=6.0",
    "rich>=12.6.0",
    "shellingham>=1.5.0.post1",
    "typer>=0.7.0",
]

requires-python = ">=3.6"

[project.urls]
Homepage = "https://github.com/MayasMess/structor"
